+++ .gitignore
@@ -13,3 +13,4 @@ node_modules/
 npm-debug.log*
 yarn-debug.log*
 yarn-error.log*
+server/public/images/*.png
